Configuration service.
Add configuration service available on port 9999. All other services are
now using it.

Improve doc to explain how to test current use cases.
