+++ authorizing/pom.xml
@@ -11,8 +11,7 @@
 
     <artifactId>authorizing</artifactId>
     <packaging>war</packaging>
-
-    <name>authorizing webapp</name>
+    <name>GeoNetwork authorizing app</name>
     <url>http://geonetwork-opensource.org</url>
 
     <properties>
@@ -27,6 +26,14 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-security</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-config</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-context</artifactId>
         </dependency>
         <dependency>
             <groupId>io.jsonwebtoken</groupId>
@@ -50,7 +57,6 @@
             <artifactId>domain</artifactId>
             <version>3.11.0-SNAPSHOT</version>
         </dependency>
-
     </dependencies>
 
     <build>
