--- authorizing/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-jwt.secret=geonetwork
-server.port=8181
-server.servlet.context-path=/
