+++ common-security/pom.xml
@@ -8,12 +8,9 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>common-security</artifactId>
-
-    <name>common</name>
-    <!-- FIXME change it to the project's website -->
-    <url>http://www.example.com</url>
+    <name>GeoNetwork common security</name>
+    <url>http://geonetwork-opensource.org</url>
 
     <dependencies>
         <dependency>
