+++ configuring/README.md
@@ -1,0 +1,12 @@
+# Configuration service
+
+Start the service using:
+```
+mvn spring-boot:run
+```
+
+It uses a file storage for storing the configuration for all services.
+
+Configuration authentication is required using user `gn` with password `gn`.
+
+Configuration can be retrieved using entry point like http://localhost:9999/geonetwork-dev.json.
