+++ configuring/src/main/resources/config/geonetwork-prod.properties
@@ -1,0 +1,12 @@
+jwt.secret=geonetwork
+
+gn.datadir=/tmp/gn-datadir
+gn.index.url=http://localhost:9200
+
+spring.datasource.url = jdbc:postgresql://localhost:5432/gn
+spring.datasource.username = www-data
+spring.datasource.password = www-data
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.database-platform = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto = update
