+++ configuring/src/test/java/org/geonetworkopensource/gnconfigserver/GnConfigServerApplicationTests.java
@@ -1,0 +1,11 @@
+package org.geonetworkopensource.gnconfigserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GnConfigServerApplicationTests {
+	@Test
+	void contextLoads() {
+	}
+}
