+++ routing/README.md
@@ -1,0 +1,6 @@
+# Routing service
+
+Start the service using:
+```
+mvn spring-boot:run
+```
