+++ routing/pom.xml
@@ -8,28 +8,13 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>routing</artifactId>
     <packaging>war</packaging>
-
-    <name>routing webapp</name>
+    <name>GeoNetwork routing webapp</name>
     <url>http://geonetwork-opensource.org</url>
 
     <properties>
-
     </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-dependencies</artifactId>
-                <version>Greenwich.SR2</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
 
     <dependencies>
         <dependency>
