--- routing/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-jwt.secret=geonetwork
-server.port=8080
-server.servlet.context-path=/
