+++ searching/README.md
@@ -1,0 +1,8 @@
+# Search service
+
+Start the service using:
+```
+mvn spring-boot:run
+```
+
+For now only valid the token created by the Authorizing app.
