+++ searching/src/main/resources/bootstrap.yml
@@ -1,0 +1,13 @@
+spring:
+  application:
+    name: gn-authorizing-service
+  profiles:
+    active: dev
+  cloud:
+    config:
+      uri: http://localhost:9999
+      name: geonetwork
+      username: gn
+      password: gn
+server:
+  port: 9990
